added splash file that contains splash widget that  will appear for 5 seconds. Edited the entry screen so it does not have the back button and appbar now. Edited the continue as a guest text in the entry screen so it navigates to the home screen
